+++ app/src/components/market/sections/common_sections/messages_text/market_resolution_message/img/Lost.tsx
