+++ app/src/components/market/sections/common_sections/messages_text/warning_message/index.tsx
