+++ app/src/components/market/sections/market_buy/market_buy.tsx
@@ -47,7 +47,7 @@ import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError } from '../../common/common_styled'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
 import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
-import { WarningMessage } from '../common_sections/messages/warning_message'
+import { WarningMessage } from '../common_sections/messages_text/warning_message'
 import { OutcomeTable } from '../common_sections/tables/outcome_table'
 import { CurrencySelector } from '../common_sections/user_transactions_tokens/currency_selector'
 import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction/transaction_details_card'
