+++ app/src/components/market/sections/market_pooling/scalar_market_pool_liquidity.tsx
@@ -52,12 +52,12 @@ import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
 import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
-import { SwitchTransactionToken } from '../../common/switch_transaction_token'
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
 import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
 import { MarketScale } from '../common_sections/card_bottom_details/scalar/market_scale'
-import { WarningMessage } from '../common_sections/messages/warning_message'
+import { WarningMessage } from '../common_sections/messages_text/warning_message'
 import { CurrencySelector } from '../common_sections/user_transactions_tokens/currency_selector'
+import { SwitchTransactionToken } from '../common_sections/user_transactions_tokens/switch_transaction_token'
 import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction/transaction_details_card'
 import { TransactionDetailsLine } from '../common_sections/user_transactions_tokens/transaction/transaction_details_line'
 import {
