+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -21,7 +21,7 @@ import { ButtonType } from '../../../../button/button_styling_types'
 import { ViewCard } from '../../../common/view_card'
 import { MarketScale } from '../../common_sections/card_bottom_details/scalar/market_scale'
 import { MarketTopDetailsOpen } from '../../common_sections/card_top_details/market_top_details_open'
-import { WarningMessage } from '../../common_sections/messages/warning_message'
+import { WarningMessage } from '../../common_sections/messages_text/warning_message'
 import { OutcomeTable } from '../../common_sections/tables/outcome_table'
 import { MarketBondContainer } from '../../market_bond/market_bond_container'
 import { MarketBuyContainer } from '../../market_buy/market_buy_container'
